VLCMedia: Don't use an eventWaiter to wait for the duration to change
The media duration is enough, and is proven to work well in the media
library.
This might break file type detection since we can't rely on the media
player, but this commit is meant to ease up the integration of libvlcpp
by removing some eventWaiter usages. We will take advantage of libvlcpp
methods after the replacement is done, instead of implementing a bunch
of methods in the VLMC's vlc binding, just so we can remove it later